updated spi drivers definitions

+++ STM32F4xx_drivers/drivers/inc/stm32f411xx_gpio_driver.h
@@ -109,7 +109,7 @@ void GPIO_ToggleOutputPin(GPIO_RegDef_t *pGPIOx, uint8_t PinNumber);
  */
 void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t ENorDI);
 void GPIO_IRQHandling(uint8_t PinNumber);
-void GPIO_PriorityConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_PriorityConfig(uint8_t IRQNumber, uint32_t IRQPriority);
 
 /*
  * Misc
